refactor schema-core document helpers
