+++ .sampo/changesets/issue-365-schema-core-documents.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/block-runtime: patch
+---
+
+Split schema-core manifest and OpenAPI document builders into a focused runtime helper module without changing the public schema-core surface.
